add mesh loading support for mz3 files
