+++ packages/niivue-desktop/src/preload/index.ts
@@ -8,6 +8,10 @@ const api = {
     const base64 = await ipcRenderer.invoke('loadFromFile', path)
     return base64
   },
+  loadStandard: async (path: string): Promise<string> => {
+    const base64 = await ipcRenderer.invoke('loadStandard', path)
+    return base64
+  },
   onToggleCrosshair: (callback: (state: boolean) => void): void => {
     ipcRenderer.on('toggleCrosshair', (_, state) => callback(state))
   }
