+++ packages/niivue-desktop/tsconfig.web.json
@@ -4,7 +4,8 @@
     "src/renderer/src/env.d.ts",
     "src/renderer/src/**/*",
     "src/renderer/src/**/*.tsx",
-    "src/preload/*.d.ts"
+    "src/preload/*.d.ts",
+    "src/common/**/*"
   ],
   "compilerOptions": {
     "composite": true,
